--- src/test/java/uk/gov/hmcts/reform/roleassignment/domain/service/common/PrepareResponseServiceTest.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.common;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class PrepareResponseServiceTest {
-
-}
