--- src/test/java/uk/gov/hmcts/reform/roleassignment/domain/service/createroles/CreateRoleAssignmentsTest.java
@@ -1,5 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.createroles;
-
-class CreateRoleAssignmentsTest {
-
-}
