--- src/test/java/uk/gov/hmcts/reform/roleassignment/domain/service/deleteroles/DeleteRoleAssignmentsTest.java
@@ -1,5 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.deleteroles;
-
-class DeleteRoleAssignmentsTest {
-
-}
