--- src/test/java/uk/gov/hmcts/reform/roleassignment/domain/service/validateroles/ValidateRoleAssignmentOperationTest.java
@@ -1,5 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.validateroles;
-
-class ValidateRoleAssignmentOperationTest {
-
-}
